work around rendering bug by changing source order
Having the inline style element that sets up the hower styles for list
elements at the start of the template triggers what must be a rendering
bug in Chromium causing the popup to be absolutely positioned with its'
top left corner at the center of the browser window instead of being
fully centered. Moving the same inline style to the end of the template
avoids triggering this issue.