include prebuilt userscript in repo
In order to address the concern about the availability of a pure
userscript version, I have added the built script to the repository.

Also to handle the interm while people still have their old scripts with
the auto-update URL set the the old address, I have added a copy in
the old location of the new (minified) build. The full non-minified
version is also available to link people to view the source and hack on
it if they like. The auto update URL will begin checking the new
location and this copy can eventually be removed.

The version has been bumped so if this lands in master with a straight
merge it is ready to go. In the future the source can be hacked and the
changes commited to master without affecting the released version. When
a release is made (and the build prop changed) the script in /dist can
be added and committed (and a tag added on the master repo).