devel branch conveniences (revert before merge!)
This is to keep git from yelling at us about changed but uncommitted
files it the repo all the time while mucking around it the devel branch.

The finished builds are not committed as part of the devel branch. It
order to bump the version number and cut an actual release, a release
branch should be made. A commit should be made to it changing the
version in build.properties, then the build script should be run. The
resulting prebuilt userscript versions should me added with `git add -f`
and committed. Only then should that release branch be merged to master
(never directly from devel).