remove interm auto-update hack file
This should not need to be present in the next release. Hopefully by
then anybody that installed the 1.3.1 release with the broken updater
will have already gotten at least gotten to here and moved on. The
remaining hack in the dist/ version of the script can be removed next
cycle, meaning after 1.3.4 or greater has hit the streets.