bump version to fix Chrome Store issue
Per discussion in #9, version 1.3.3 git incorrectly uploaded
to the Chrome Store with the userscript code instead ef the
extension code. This is a version bump so that the package can
be re-uploaded and existing users will not get and error next
time the userscript version is update.

Rebuilt packages to include hotfix for URL issue from 1fad101