remove release notes from readme; fix credits
The release notes section of the readme is duplicated with the data it
tags which can be browsed separately. Removed section in favor of just a
credits section with consistently formatted links to SE profiles, Github
commits (if present) ordered by release number of first contrubution.

The issues/plans section is also obsoleted by having all that data it
Github, so remove that too.