revamp site matching rules to work across browsers
The userscript pattern matching is much more tollerant that what
browsers expect. What regex is allowed varries between FF and Chrome.
This set of rules seems to work across the board with the one exception
that Chrome does not treat '.' as a wild-card like must of the other
systems, so there is a rule for fixing that at build time.

Also removed the OnStartup site as it was closed by SE.